Copy non-markdown files

+++ docgen.py
@@ -10,19 +10,20 @@ markdowner = Markdown(extras=["tables", "fenced-code-blocks"])
 
 
 class Document():
-    src_file: str  # overview.md
+    src_filename: str  # overview.md
+    src_path: str  # raw/technical/tools/overview.md
     name: str  # overview
     extension: str  # .md
-    path: str  # technical/tools
+    dir_path: str  # technical/tools
     markdown: str
 
-    def __init__(self, src_file, path, markdown=""):
-        self.src_file = src_file
-        (name, extension) = os.path.splitext(src_file)
+    def __init__(self, src_filename, src_path, dir_path):
+        self.src_filename = src_filename
+        self.src_path = src_path
+        (name, extension) = os.path.splitext(src_filename)
         self.name = name.replace(" ", "-").lower()
         self.extension = extension
-        self.path = path
-        self.markdown = markdown
+        self.dir_path = dir_path
 
     def is_doc(self) -> bool:
         return self.extension == ".md"
@@ -43,7 +44,8 @@ def create_styles() -> str:
 
 
 def cleanup():
-    shutil.rmtree(OUTPUT_DIR)
+    if os.path.exists(OUTPUT_DIR):
+        shutil.rmtree(OUTPUT_DIR)
 
 
 # TODO: Replace <title> with content filename
@@ -59,8 +61,9 @@ def walk_raw_docs() -> [Document]:
     for (root, _, filenames) in os.walk(RAW_DIR):
         for filename in filenames:
             dirpath = os.path.relpath(root, RAW_DIR).replace(".", "")
+            raw_path = os.path.join(root, filename)
 
-            doc = Document(filename, dirpath)
+            doc = Document(filename, raw_path, dirpath)
             if doc.is_doc():
                 with open(os.path.join(root, filename), 'r') as doc_file:
                     markdown = doc_file.read().strip()
@@ -76,15 +79,22 @@ def generate_docs(raw_docs):
     styles = create_styles()
 
     for doc in raw_docs:
-        page = build_doc_page(base, styles, doc.markdown)
-        doc_dir_path = os.path.join(OUTPUT_DIR, doc.path)
-        doc_path = os.path.join(doc_dir_path, f"{doc.name}.html")
-
+        doc_dir_path = os.path.join(OUTPUT_DIR, doc.dir_path)
         if not os.path.exists(doc_dir_path):
             os.makedirs(doc_dir_path)
 
-        with open(doc_path, 'w') as output_file:
-            output_file.write(page)
+        if doc.is_doc():
+            page = build_doc_page(base, styles, doc.markdown)
+
+            doc_path = os.path.join(doc_dir_path, f"{doc.name}.html")
+
+            with open(doc_path, 'w') as output_file:
+                output_file.write(page)
+        else:
+            # raw file copy
+            src = doc.src_path
+            target = os.path.join(doc_dir_path, doc.src_filename)
+            shutil.copyfile(src, target)
 
 
 if __name__ == "__main__":
@@ -96,4 +106,4 @@ if __name__ == "__main__":
     print(f"Found {len(raw_docs)} files to generate")
 
     generate_docs(raw_docs)
-    print("Generated documents")
+    print("All documents generated")
